+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDProfileId.cs
@@ -104,6 +104,7 @@ namespace UnityEngine.Rendering.HighDefinition
         RaytracingIntegrateIndirectDiffuse,
         RaytracingFilterIndirectDiffuse,
         RaytracingDebugOverlay,
+        RayTracingRecursiveRendering,
 
         // Profile sampler for prepare light for GPU
         PrepareLightsForGPU,